python/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-text-splitters==0.3.1
langcodes==3.4.1
langsmith==0.1.138
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.53.0
orjson==3.10.10
packaging==24.1
preshed==3.0.9
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
setuptools==75.3.0
shellingham==1.5.4
smart-open==7.0.5
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
tenacity==9.0.0
tqdm==4.66.6
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yarl==1.17.1
